server/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.7
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.12.2
future==0.18.3
gevent==23.9.1
greenlet==3.0.1
idna==3.4
platformdirs==3.10.0
pycparser==2.21
PySocks==1.7.1
pytz==2023.3.post1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.4
tldextract==5.1.1
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.2
zope.event==5.0
zope.interface==6.1
